Revert "chore: cherry-pick changes to 4.25.1 stable (#4054)"
This reverts commit 0b200fd53b608197cc45c551d0f0558018094ada.
