+++ android/src/main/java/com/swmansion/rnscreens/ScreenStackFragment.kt
@@ -270,14 +270,6 @@ class ScreenStackFragment :
                 object : WindowInsetsAnimationCompat.Callback(
                     WindowInsetsAnimationCompat.Callback.DISPATCH_MODE_STOP,
                 ) {
-                    override fun onPrepare(animation: WindowInsetsAnimationCompat) {
-                        super.onPrepare(animation)
-
-                        if ((animation.typeMask and WindowInsetsCompat.Type.ime()) != 0) {
-                            sheetDelegate.notifyKeyboardAnimationStart()
-                        }
-                    }
-
                     // Replace InsetsAnimationCallback created by BottomSheetBehavior
                     // to avoid interfering with custom animations.
                     // See: https://github.com/software-mansion/react-native-screens/pull/2909
@@ -295,10 +287,6 @@ class ScreenStackFragment :
 
                     override fun onEnd(animation: WindowInsetsAnimationCompat) {
                         super.onEnd(animation)
-
-                        if ((animation.typeMask and WindowInsetsCompat.Type.ime()) != 0) {
-                            sheetDelegate.notifyKeyboardAnimationEnd()
-                        }
 
                         screen.onSheetYTranslationChanged()
                     }
