+++ apps/src/tests/issue-tests/index.ts
@@ -191,7 +191,6 @@ export { default as Test3833 } from './Test3833';
 export { default as Test3835 } from './Test3835';
 export { default as Test3867 } from './Test3867';
 export { default as Test3885 } from './Test3885';
-export { default as Test4027 } from './Test4027';
 export { default as TestScreenAnimation } from './TestScreenAnimation';
 // The following test was meant to demo the "go back" gesture using Reanimated
 // but the associated PR in react-navigation is currently put on hold
